Some fixes and improvements
Reading a blank value from FST_Config will no longer close out the cmd
(even though this shouldn't occur outside of a user edit)
Writing the server config sometimes fails, still tryint to figure out
why, but if it does we will now warn the user and not continue.
If the server shutsdown because of an error create a way to dump all
config values into one file with the error log so i don't have to ask
people to upload 4 files.